Fix: error handling on relay version check
If we have a network error while performing the version check between
the session daemon and the relay, we should not report to the user that
there is a version mismatch. We now return LTTNG_ERR_RELAYD_VERSION_FAIL
on relayd_version_check() when the daemons are not compatible and the
negative value from sendmsg/recvmsg on network errors.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>