Fix: reply to version check even on protocol mismatch
In the relay, we currently put the connection when we detect the major
version from the session daemon is not compatible. We don't reply to the
version check message. The relay still holds a reference on the
connection so it is not closed and the session daemon is left blocking
in recvmsg.

The relay now replies to the version check so the session daemon knows
it is not compatible, and the relay completely closes the connection on
its side and removes the FD from the poll set.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>