Cleanup: misleading variable name
Using "running" implies that the thread is guaranteed to be
functional/ready. The intention of those "running" flags is only
to indicate that the underlying pthread was created. The thread
may not be running anymore and these flags should not be used
to check if the thread is "ready" to process anything.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>