Fix: sessions with agent channels fail to load
Channels of the "agent" types cannot be created directly. They are
meant to be created implicitly through the activation of events in
their domain.

However, a user can override the default channel configuration
attributes by creating the underlying UST channel before enabling an
agent domain event.

Hence, the channel's type is substituted before the creation and
restored by the time events are created.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>