Clean-up: explicit mb before decrementing lttng_sessiond_ready
This is mostly a documentation fix as there are no thread-safety
implications to this change. uatomic_sub_return() was used since it
performs a full memory barrier before and after the atomic operation
(as per the urcu documentation).

The barrier performed after the substraction is not needed in this
particular case. Moreover, using an explicit cmm_smp_mb() statement
makes the code clearer; see the comment as to why this barrier is
needed.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>