Use dynamic buffer to build session configuration path
Re-use the dynamic buffer interface to build the session
configuration path. The main benefit here is silencing a source
string truncation warning emitted by GCC 8. However, this
interface is also simpler to use than manually building the
path.

The LTTNG_PATH_MAX is still enforced, but there is no real
need to restrict paths to that size. This could be removed if
it ever poses a problem.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>