Add unused attribute to lttng_to_index_major param
Suppresses an unused variable warning. The parameter is kept
since this does depend on the connection's full version. the
'minor' parameter is unused for now since there is only one
major version to support and only one major index version.

However, this could change in the future and we don't want to
have to modify all the version conversion sites.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>