Perform relayd socket pair cleanup on control socket transmission error
A reference to the local ctx for the socket pair is used to "force" an
evaluation of the data and metadata streams since we changed the endpoint
status. This imitate what is currently done for the data socket.

This prevent hitting network timeouts multiple time. For now, there is
no mechanism for retry hence "terminating" all communication make sense
and prevent unwanted delays on operation.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>