rotation-api: introduce rotation schedule descriptors
The current rotation API is limited in that it imposes two types
of automatic rotation. This change introduces rotation schedules,
a more generic concept, that will eventually allow an arbitrary
number of rotation schedules to be set on a session.

Moreover, the API's design is now more aligned with the APIs
that were recently added (i.e. use of opaque descriptors for
commands that may require more parameters at some point).

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>