Fix: consumers don't honor protocol on rotate pending relay command
The payload of a command should not be sent to the sessiond if
its execution failed. The sessiond reports an error as soon as
the return code of the command is returned and will never consume
the payload, resulting in a protocol error down the line.

Moreover, the ust-consumer's implementation doesn't check for
errors as it stores the return value of
lttng_consumer_rotate_pending_relay() to an unsigned variable and
checks for negative values.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>