Use the dynamic buffer to serialize notification objects
The objects of the notification/trigger APIs are currently
serialized by calling their serialization functions with a NULL
buffer, returning the expected size, and then calling the
serialization function with a sufficiently-large buffer.

This code predates the dynamic buffer utility which can now be
used to perform serialization operations in one pass.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>